# docs/swagger.yaml
info:
  contact: {}
paths:
...
  /project/addProject:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 项目名称
        in: formData
        name: name
        type: string
      - description: 项目 URL
        in: formData
        name: url
        type: string
      - description: 项目描述
        in: formData
        name: describe
        type: string
      - description: 项目描述1
        in: formData
        name: describe1
        type: string
      - description: 项目图片1
        in: formData
        name: image1
        type: file
      - description: 项目描述2
        in: formData
        name: describe2
        type: string
      - description: 项目图片2
        in: formData
        name: image2
        type: file
      - description: 项目描述3
        in: formData
        name: describe3
        type: string
      - description: 项目图片3
        in: formData
        name: image3
        type: file
      - description: 项目描述4
        in: formData
        name: describe4
        type: string
      - description: 项目图片4
        in: formData
        name: image4
        type: file
      - description: 项目描述5
        in: formData
        name: describe5
        type: string
      - description: 项目图片5
        in: formData
        name: image5
        type: file
      - description: 项目描述6
        in: formData
        name: describe6
        type: string
      - description: 项目图片6
        in: formData
        name: image6
        type: file
      - description: 项目描述7
        in: formData
        name: describe7
        type: string
      - description: 项目图片7
        in: formData
        name: image7
        type: file
      - description: 项目描述8
        in: formData
        name: describe8
        type: string
      - description: 项目图片8
        in: formData
        name: image8
        type: file
      - description: 项目描述9
        in: formData
        name: describe9
        type: string
      - description: 项目图片9
        in: formData
        name: image9
        type: file
      - description: 项目描述10
        in: formData
        name: describe10
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 添加新的项目
      tags:
      - 项目模块
  /project/deleteProject:
    post:
      parameters:
      - description: 项目ID
        in: query
        name: id
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 删除项目
      tags:
      - 项目模块
...
  /project/getTheProject:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 项目ID
        in: formData
        name: id
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 获取指定的项目
      tags:
      - 项目模块
  /project/updateProject:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 项目ID
        in: formData
        name: id
        type: string
      - description: 项目名称
        in: formData
        name: name
        type: string
      - description: 项目 URL
        in: formData
        name: url
        type: string
      - description: 项目描述
        in: formData
        name: describe
        type: string
      - description: 项目描述1
        in: formData
        name: describe1
        type: string
      - description: 项目图片1
        in: formData
        name: image1
        type: file
      - description: 项目描述2
        in: formData
        name: describe2
        type: string
      - description: 项目图片2
        in: formData
        name: image2
        type: file
      - description: 项目描述3
        in: formData
        name: describe3
        type: string
      - description: 项目图片3
        in: formData
        name: image3
        type: file
      - description: 项目描述4
        in: formData
        name: describe4
        type: string
      - description: 项目图片4
        in: formData
        name: image4
        type: file
      - description: 项目描述5
        in: formData
        name: describe5
        type: string
      - description: 项目图片5
        in: formData
        name: image5
        type: file
      - description: 项目描述6
        in: formData
        name: describe6
        type: string
      - description: 项目图片6
        in: formData
        name: image6
        type: file
      - description: 项目描述7
        in: formData
        name: describe7
        type: string
      - description: 项目图片7
        in: formData
        name: image7
        type: file
      - description: 项目描述8
        in: formData
        name: describe8
        type: string
      - description: 项目图片8
        in: formData
        name: image8
        type: file
      - description: 项目描述9
        in: formData
        name: describe9
        type: string
      - description: 项目图片9
        in: formData
        name: image9
        type: file
      - description: 项目描述10
        in: formData
        name: describe10
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 修改指定的项目
      tags:
      - 项目模块
  /skill/addSkill:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 技能名字
        in: formData
        name: name
        type: string
      - description: 技能网址
        in: formData
        name: url
        type: string
      - description: 技能描述
        in: formData
        name: description
        type: string
      - description: 技能时间
        in: formData
        name: time
        type: string
      - description: 技能图片
        in: formData
        name: image
        type: file
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 添加新的技能
      tags:
      - 技能模块
...
  /visitor/check:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 请输入Token
        in: formData
        name: token
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 查询访问者角色
      tags:
      - 访问者模块
  /visitor/findVisitor:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 请输入用户名
        in: formData
        name: username
        type: string
      - description: 请输入密码
        in: formData
        name: password
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 查询访问者信息
      tags:
      - 访问者模块
  /visitor/registerVisitor:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 请输入用户名
        in: formData
        name: username
        type: string
      - description: 请输入密码
        in: formData
        name: password
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 注册
      tags:
      - 访问者模块
swagger: "2.0"